fix(delivery): rename mislabeled VPS audit metric to cumulative_pushed_secs (#232)
The DiskCachePushSample audit payload carried a field `current_chunk_delay_secs`
whose value on the rust-push path was `duration_processed_ms / 1000.0` —
CUMULATIVE media pushed (≈ stream age), NOT behind-live lag. It surfaced ~7800s
in the audit log and read like a behind-live number, misleading operators.

No host-side or frontend parser reads this field (grep: only emitted into the
audit JSON), so the honest, lowest-risk fix is to rename the key to reflect
what it is.

- disk_cache_push_sample::emit_push_sample (production emitter): rename param +
  JSON key to cumulative_pushed_secs; document that the behind-live signal is
  chunk_supply_lag_ms / the dashboard's per-endpoint chunk_delay_secs.
- endpoint_task consumer_task: both pusher paths now pass the cumulative value
  under the new name. The ffmpeg path previously passed per-chunk duration —
  even more wrong under the old name — now reads duration_processed_ms under the
  same lock so both paths mean the same thing.
- endpoint_reader::build_push_sample_payload (test-only helper) + its tests:
  rename to match.
- audit.rs doc comment updated.

Also extract build_rtmp_url + its test re-export into endpoint_rtmp_url.rs
(#[path] submodule, the established cap-dodging pattern in this crate). The FIX A
+ FIX C additions pushed endpoint_task.rs to 1016 lines (it was already at 999);
the extraction brings it back to 985, logic-neutral. build_rtmp_url_pub is
re-exported at the endpoint_task module level so super::super::build_rtmp_url_pub
in the tests still resolves.

Refs #232
