+++ crates/rs-delivery/src/disk_cache/endpoint_reader.rs
@@ -150,7 +150,7 @@ fn build_push_sample_payload(
     inter_chunk_gap_ms: u64,
     chunk_duration_ms: u64,
     delivery_delay_secs: u64,
-    current_chunk_delay_secs: f64,
+    cumulative_pushed_secs: f64,
 ) -> serde_json::Value {
     let burst_factor = if inter_chunk_gap_ms == 0 {
         0.0
@@ -164,7 +164,9 @@ fn build_push_sample_payload(
         "inter_chunk_gap_ms": inter_chunk_gap_ms,
         "burst_factor": burst_factor,
         "delivery_delay_secs": delivery_delay_secs,
-        "current_chunk_delay_secs": current_chunk_delay_secs,
+        // Total media pushed so far (≈ stream age), NOT behind-live lag. See
+        // disk_cache_push_sample::emit_push_sample for the rename rationale.
+        "cumulative_pushed_secs": cumulative_pushed_secs,
     })
 }
 
@@ -280,7 +282,7 @@ mod tests {
             /* inter_chunk_gap_ms = */ 850,
             /* chunk_duration_ms = */ 1000,
             /* delivery_delay_secs = */ 120,
-            /* current_chunk_delay_secs = */ 151.3,
+            /* cumulative_pushed_secs = */ 151.3,
         );
         assert_eq!(payload["endpoint"], "FB-NewLevel");
         assert_eq!(payload["chunk_id"], 100);
@@ -289,7 +291,7 @@ mod tests {
         let burst = payload["burst_factor"].as_f64().unwrap();
         assert!((burst - (1000.0 / 850.0)).abs() < 1e-6);
         assert_eq!(payload["delivery_delay_secs"], 120);
-        let cd = payload["current_chunk_delay_secs"].as_f64().unwrap();
+        let cd = payload["cumulative_pushed_secs"].as_f64().unwrap();
         assert!((cd - 151.3).abs() < 1e-6);
     }
